LayoutTests/http/tests/resourceLoadStatistics/user-interaction-in-cross-origin-sub-frame-expected.txt

Tests that user interaction is always logged for the top document/frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.location.origin is topFrameOrigin
PASS testRunner.isStatisticsHasHadUserInteraction(topFrameOrigin) is false
PASS testRunner.isStatisticsHasHadUserInteraction(subFrameOrigin) is false
PASS testRunner.isStatisticsHasHadUserInteraction(topFrameOrigin) is true
PASS testRunner.isStatisticsHasHadUserInteraction(subFrameOrigin) is false
PASS successfullyParsed is true

TEST COMPLETE
